ECO-269: rename VestingHash to NamedHash, and rename path of Vesting page.
